templates/requirements.txt
cocoindex
python-dotenv
psycopg[binary]
psycopg-pool
mcp
pyyaml
sentence-transformers
